move test env stuff into new file
